HSEARCH-2065 Reorganizes the code to fix visibility issues
