+++ elasticsearch/src/main/java/org/hibernate/search/backend/elasticsearch/impl/ElasticsearchLuceneQueryTranslator.java
@@ -10,8 +10,6 @@ import java.util.Properties;
 import java.util.Set;
 
 import org.apache.lucene.search.Query;
-import org.hibernate.search.backend.elasticsearch.ElasticsearchQueries;
-import org.hibernate.search.backend.elasticsearch.json.ToElasticsearch;
 import org.hibernate.search.engine.integration.impl.ExtendedSearchIntegrator;
 import org.hibernate.search.engine.service.spi.Startable;
 import org.hibernate.search.engine.spi.EntityIndexBinding;
@@ -46,7 +44,7 @@ public class ElasticsearchLuceneQueryTranslator implements LuceneQueryTranslator
 		JsonObject query = new JsonObject();
 		query.add( "query", convertedQuery );
 
-		return ElasticsearchQueries.fromJson( query );
+		return new ElasticsearchJsonQueryDescriptor( query );
 	}
 
 	@Override
